+++ test/mailers/user_mailer_test.rb
@@ -74,4 +74,15 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/You have a new followup survey waiting to be completed on MyApnea\.Org\./, email.encoded)
   end
 
+  test "encounter digest email" do
+    owner = users(:owner)
+
+    email = UserMailer.encounter_digest(owner, 48, {}).deliver_now
+    assert !ActionMailer::Base.deliveries.empty?
+
+    assert_equal [users(:owner).email], email.to
+    assert_equal "48 Followup Surveys Launched on #{Date.today.strftime('%a %d %b %Y')}", email.subject
+    assert_match(/Today, 48 surveys were launched to members of MyApnea\.Org\./, email.encoded)
+  end
+
 end